requirements.txt
flask>=3.0.0
groq>=0.9.0
yt-dlp>=2024.1.0
openai-whisper>=20231117
playwright>=1.44.0